Tools: aero_upload: Give some time to kill and start mavlink-router
When running for the first time it was always failling and printing:
echo "Process \$p is running and keeping UART busy"
because the mavlink-routerd was still being killed on the if.
A similar problem was hapenning then starting mavlink-router,
the ssh command line was being exit before the mavlink-router was
initialized.

This 2 sleeps looks hackish but they work and we should not spend
so much time in a so small script that if it failed we can run again
to fix.