systemcmds: Add aerofc_fpga_config
This command will configure Aero FPGA and PX4 internals to match
AEROFC_I2C_TELEM and AEROFC_UART_INV parameters.

AEROFC_I2C_TELEM will turn off the UART in the telemetry connector
and turn on the I2C in the same connector, so user can choose
if the telemetry connector will be used to connect a external sensor
like a lidar or not. Because of the limitation in the exported lines
from microcontroller AeroFC can not have both.

AEROFC_UART_INV will invert the polarity of the UART, this is userful
when connecting a SBUS receiver.